+++ src/util.rs
@@ -2,7 +2,7 @@ use bevy::prelude::*;
 use crate::{configs::GL, pos::Pos};
 // 棋盘上交点的很纵坐标 (horizon 0..=8, verticle 0..=9)
 // return the middle xy cordinate of the grid cross point
-pub fn grid2xy(horizon: u8, verticle: u8) -> Vec2 {
+pub fn grid2xy(horizon: i32, verticle: i32) -> Vec2 {
     assert!(horizon <= 8 && verticle <= 9);
     return Vec2::new(horizon as f32 * GL - 4.0 * GL, verticle as f32 * GL - 4.5 * GL);
 }
@@ -19,8 +19,8 @@ pub fn in_bound(world_position: &Vec2) -> bool {
 
 pub fn world_to_board(world_position: &Vec2) -> Pos {
     return Pos {
-        0: ((world_position.x + 4.5 * GL) / GL) as u8,
-        1: ((world_position.y + 5.0 * GL) / GL) as u8,
+        0: ((world_position.x + 4.5 * GL) / GL) as i32,
+        1: ((world_position.y + 5.0 * GL) / GL) as i32,
     }
 }
